src/requirements.txt
caldav==1.3.9
Cython==3.0.8
Quart==0.19.4
gunicorn==21.2.0
uvicorn[standard]==0.26.0
icalendar==5.0.11
mattermostautodriver==2.0.0
asyncpg==0.29.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
pytest==7.4.4
aiohttp==3.9.1
APScheduler==4.0.0a4
pytest-asyncio==0.23.4a2
coverage==7.4.0
dramatiq[rabbitmq]==1.15.0
alembic==1.13.1
